+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.0] Updated vJoy version
+- Added: Support for up to 4 continuous hat switches in the range of 0..360° with a 1/100° resolution.
+- Changed: Max number of axis increased to 16
+- Changed: Supported vJoy driver version updated to 2.2.2.1 from: https://github.com/njz3/vJoy/
+
 ## [0.2.0] Hat switch support
 - Added: Support for up to four 4-way hat switches (e.g. a D-pad).